requirements.txt
#pip install -r requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils
python-jose
mysqlclient
python-dotenv
pandas